Added rough allocation function for future use
Was going to try and use these memory allocation functions now but will hold off until I have a definite need for them. So instead of trying to 'decide' ahead of time what features to start adding I will try and let features naturally come about when there is a need for them. See if that is a better way to do these things for now.