Fixed bug - now have single image showing instead of entire png
Since spine works with an atlas you first load in the entire packed png file which contains all your images for your character in that pose. I was only able to display this image in it's entirety in the previous commit. I know figured out how to display only the image I wanted to display such as "rear-bracer".