Added SpineFuncPtrTest to GameState to hopefully fix bug
When working with current code I was still getting a bug with dll reloading while input playback was enabled. I think this might of had something to do with how global variables are stored upon dll reload so I just moved the Spine func ptr to gamestate and that seems to have 'fixed' the issue.