Have basic minkowski like collision detection
Able to get collision reporting based on minkowski summing of shapes. Things to keep in mind: Spine's bounding boxes (which I create in spine's editor and use for collision detection) are stateless and are shared between skeletons. So you can't grab an individual skeleton's bounding box. Instead you use 'spVertex_computeWorldVertices' to get world vertices which are stateful and depend on skeleton.