Have begun to form the idea of how hit boxes will work
In order to separate collision detection from animation I will be generating hit boxes in certain positions, depending on the animation playing, over a specified period of time. So a punch animation's hitBox might display .4 seconds in to its timer being started at pos 400, 500. This code is the first step towards that