Fixed memory handling, dynam allocator logic
When I was releasing a block of memory through the memory_handling logic, I wasn't properly handling the situation where the block being released was the last block in the chain. Essentially the size of the memory block's header wasn't being taken into account when freeing. Also I wasn't properly checking to see which block was the last block.