Implemented a RenderToImage func
This mimics opengl's "RenderToTexture()" function where the user can pre-render an image and save it for later rendering to color buffer. This can be useful in situations where the user might want perform some calculations to determine a final rendered image or generate a composite image of multiple images at runtime.