Hacked on a new convention to enforce out params
Since it was bothering me how thngs are currently handled in c++ with out parameters and return values, I decided to have a sort of compiler enforced out parameter that you can specifiy with a r-value reference for a function. This makes it easier to see when a function will be modifying some parameters, making them essentially return values. Right now I would say anytime I see a return value that's because that is a new variable that was not previously created by me.