Updated DrawTextureSlowly to compile again.
Currently I'm trying to figure out how to implement a new feature into renderer so I need to walk though how things are currently being draw. Updated DrawTextureSlowly so that I can use it to help understand all that's going on when rendering pixels to the screen (eaiser with DrawTextureSlowly function since it is unoptimized).