Changed atlas code to take region's pos as bottomleft instead topleft
Spine originally identified the x,y postition of an image region (atlas region) within an atlas page as the top left of the image. It also was thinking in terms of top down instead of bottom up in regards to y position. Changed a calcuation so it is instaed bottomleft y position and bottom up.