Fixed bug concerning overwriting wrong portion of framebuffer
When a character reached the end of the screen some of their body parts would overflow and be drawn to the beginning of the screen. This had to do with 8 pixel wide instructions and some of the 8 wide pixel structure overflowing the bounds of the screen. Fixed this by adding a clip mask to clip all oiverflowed pixels and just keep what was in dest pixel before.