Switched gears to working on Hash map/table
Got to a point where std lib kept crashing when trying to use a standard 'unordered_map' container from it (cause c++ is awful). I tired a few other C libs for this but wasn't thrilled with them so just implementing my own. Currently have a 'working' implementation with plenty of holes.