Fixed bug with UpdateAnimationState logic
With the way things were setup before I opended the possibility up for wierd bugs due to the scenario where, between the prev frame and current frame, there was still some time left over for the animaiton to play out. This was left over because in prev frame the current anim runtime was less than the anim total runtime and in current frame the runtime of anim jumped enough to surpass the total anim runtime. This left so many ms of time left for anim to still play out. Added quick fix here