Might have fixed lighting bug
Dont think original lighting equation was correctly testing when normal angle was between the threshold angle values. I think this new one does the trick. Basically I'm taking the original light vector angle (say 1 radian) and adding and subtracting that angle by the threshold value so I get a range of angles (e.g. threshold value = 1 so 1+1 = 2 and 1-1 = 0. So range is from 0 radians to 2 radians). Then I just check to see, in counter-clockwise fashion, if the normal angle from normal map falls within the range between 0 and 2. If it does then the pixel is lit. Still might be bugs in this but at least making headway