Have both world and camera transforms refactored
Right now, I have refactored the code to have separate world and camera basis functions but the same transform function. For me, this makes things easier to understand as I'm just creating basis vectors for each stage that will then be used to transform the child space vector into it's parent space vector. I'm going to try and do the same with projection (so I'll have a function that produces a projection space basis).