So, it seems maybe my actual camera transform is correct?
So I seem to be getting the same behavior out of both glm and my code for camera transforms. If there is an issue it might be with how I am currently trying to lock a world position to look at. I might try and add same different camera controls and see how things look for both implementations.